+++ mycodo/inputs/bme280_rpi_bme280.py
@@ -5,6 +5,7 @@ from mycodo.inputs.base_input import AbstractInput
 from mycodo.inputs.sensorutils import calculate_altitude
 from mycodo.inputs.sensorutils import calculate_dewpoint
 from mycodo.inputs.sensorutils import calculate_vapor_pressure_deficit
+from mycodo.inputs.sensorutils import convert_from_x_to_y_unit
 
 # Measurements
 measurements_dict = {
@@ -107,9 +108,14 @@ class InputModule(AbstractInput):
 
         data = self.sensor.sample(self.bus, self.i2c_address, self.calibration_params)
 
-        self.value_set(0, data.temperature)
-        self.value_set(1, data.humidity)
-        self.value_set(2, data.pressure)
+        if self.is_enabled(0):
+            self.value_set(0, data.temperature)
+
+        if self.is_enabled(1):
+            self.value_set(1, data.humidity)
+
+        if self.is_enabled(2):
+            self.value_set(2, convert_from_x_to_y_unit('hPa', 'Pa', data.pressure))
 
         if self.is_enabled(0) and self.is_enabled(1):
             self.value_set(3, calculate_dewpoint(self.value_get(0), self.value_get(1)))